Fixed some riscv-tests

+++ sim/common/rvfloats.h
@@ -27,13 +27,13 @@ uint32_t rv_ltof_s(uint64_t a, uint32_t frm, uint32_t* fflags);
 uint32_t rv_lutof_s(uint64_t a, uint32_t frm, uint32_t* fflags);
 
 uint32_t rv_fclss_s(uint32_t a);
-uint32_t rv_fsgnj_s(uint32_t a, uint32_t b);
-uint32_t rv_fsgnjn_s(uint32_t a, uint32_t b);
-uint32_t rv_fsgnjx_s(uint32_t a, uint32_t b);
+uint32_t rv_fsgnj_s(uint64_t a, uint64_t b);
+uint32_t rv_fsgnjn_s(uint64_t a, uint64_t b);
+uint32_t rv_fsgnjx_s(uint64_t a, uint64_t b);
 
 uint32_t rv_flt_s(uint32_t a, uint32_t b, uint32_t* fflags);
 uint32_t rv_fle_s(uint32_t a, uint32_t b, uint32_t* fflags);
-uint32_t rv_feq_s(uint32_t a, uint32_t b, uint32_t* fflags);
+uint32_t rv_feq_s(uint64_t a, uint64_t b, uint32_t* fflags);
 uint32_t rv_fmin_s(uint32_t a, uint32_t b, uint32_t* fflags);
 uint32_t rv_fmax_s(uint32_t a, uint32_t b, uint32_t* fflags);
 
@@ -49,12 +49,12 @@ uint64_t rv_fmsub_d(uint64_t a, uint64_t b, uint64_t c, uint32_t frm, uint32_t* 
 uint64_t rv_fnmadd_d(uint64_t a, uint64_t b, uint64_t c, uint32_t frm, uint32_t* fflags);
 uint64_t rv_fnmsub_d(uint64_t a, uint64_t b, uint64_t c, uint32_t frm, uint32_t* fflags);
 
-uint64_t rv_ftoi_d(uint64_t a, uint64_t frm, uint32_t* fflags);
-uint64_t rv_ftou_d(uint64_t a, uint64_t frm, uint32_t* fflags);
-uint64_t rv_ftol_d(uint64_t a, uint64_t frm, uint32_t* fflags);
-uint64_t rv_ftolu_d(uint64_t a, uint64_t frm, uint32_t* fflags);
-uint64_t rv_itof_d(uint32_t a, uint32_t frm, uint32_t* fflags);
-uint64_t rv_utof_d(uint32_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_ftoi_d(uint64_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_ftou_d(uint64_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_ftol_d(uint64_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_ftolu_d(uint64_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_itof_d(uint64_t a, uint32_t frm, uint32_t* fflags);
+uint64_t rv_utof_d(uint64_t a, uint32_t frm, uint32_t* fflags);
 uint64_t rv_ltof_d(uint64_t a, uint32_t frm, uint32_t* fflags);
 uint64_t rv_lutof_d(uint64_t a, uint32_t frm, uint32_t* fflags);
 
@@ -69,7 +69,7 @@ uint64_t rv_feq_d(uint64_t a, uint64_t b, uint32_t* fflags);
 uint64_t rv_fmin_d(uint64_t a, uint64_t b, uint32_t* fflags);
 uint64_t rv_fmax_d(uint64_t a, uint64_t b, uint32_t* fflags);
 
-uint64_t rv_dtof(uint64_t a);
+uint32_t rv_dtof(uint64_t a);
 uint64_t rv_ftod(uint32_t a);
 
 #ifdef __cplusplus
